base.html 디자인 수정, 자재 창고 ETC 삭제 방지, 업체 목록, 자재 목록 디자인 수정, 자재 등록 및 수정 필수 입력 사항 문구 추가.

+++ src/main/java/com/glkids/procurehub/controller/rest/MaterialRestController.java
@@ -54,12 +54,18 @@ public class MaterialRestController {
     public ResponseEntity<MaterialWarehouseDTO> updateWarehouse(@RequestBody MaterialWarehouseDTO materialWarehouseDTO) {
         materialWarehouseDTO.setWrhscode(materialWarehouseDTO.getWrhscode().trim().toUpperCase());
         materialWarehouseDTO.setWrhsname(materialWarehouseDTO.getWrhsname().trim().toUpperCase());
+        if (materialWarehouseDTO.getWrhscode().equals("ETC")) {
+            return ResponseEntity.ok(null);
+        }
         return ResponseEntity.ok(materialService.registerMaterialWarehouse(materialWarehouseDTO));
     }
 
     @PostMapping(value = "/warehousedelete", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Boolean> deleteWarehouse(@RequestBody MaterialWarehouseDTO materialWarehouseDTO) {
         materialWarehouseDTO.setWrhscode(materialWarehouseDTO.getWrhscode().trim().toUpperCase());
+        if (materialWarehouseDTO.getWrhscode().equals("ETC")) {
+            return ResponseEntity.ok(false);
+        }
         return ResponseEntity.ok(materialService.deleteMaterialWarehouse(materialWarehouseDTO.getWrhscode()));
     }
 }
